Cast MAX_INT to long before adding 1 to test our of range

+++ src/test/java/jimmy/mcgymmy/logic/parser/ParserUtilTest.java
@@ -54,7 +54,8 @@ public class ParserUtilTest {
     @Test
     public void parseIndex_outOfRangeInput_throwsParseException() {
         assertThrows(ParseException.class,
-                Index.MESSAGE_INVALID_INDEX, () -> ParserUtil.parseIndex(Long.toString(Integer.MAX_VALUE + 1)));
+                Index.MESSAGE_INVALID_INDEX, () -> ParserUtil.parseIndex(
+                        Long.toString(((long) Integer.MAX_VALUE + 1))));
     }
 
     @Test
